oh ooohhh, mixing up types is evil! ;)
fix bug that led's only showed up in grey/black



git-svn-id: https://ktechlab.svn.sourceforge.net/svnroot/ktechlab/trunk@175 76876f85-f02d-0410-881d-fd707285c4ba

+++ src/electronics/components/led.cpp
@@ -57,9 +57,9 @@ LED::~LED()
 void LED::dataChanged()
 {
 	QColor color = dataColor("0-color");
-	r = color.red() / 0x100;
-	g = color.green() / 0x100;
-	b = color.blue() / 0x100;
+	r = color.red() / (double)0x100;
+	g = color.green() / (double)0x100;
+	b = color.blue() / (double)0x100;
 }
 
 void LED::stepNonLogic()
